requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.3
Pillow==10.2.0 